Add Comments
- Code Comments are add
- arrange the file to format
