+++ Project Euler/src/Q1/src/Q1.cpp
@@ -2,6 +2,7 @@
 
 using namespace std;
 
+// Function to find the sum of all the multiples of 3 or 5 below 1000
 int sum() {
     int sum = 0;
     for (int i = 1; i < 1000; i++) {
