+++ Project Euler/src/Q1/src/Q1.h
