+++ Project Euler/src/Q10/src/Q10.h
