+++ Project Euler/src/Q2/src/Q2.h
