+++ Project Euler/src/Q3/src/Q3.h
