+++ Project Euler/src/Q5/src/Q5.cpp
@@ -2,9 +2,11 @@
 
 using namespace std;
 
+// Function to find the smallest number that is evenly divisible by all of the numbers from 1 to 20
 int maxNumber()
 {
     long i;
+    // i is the number to be checked
     for (i = 20; ; i++) {
         int count = 0;
         for (int j = 1; j <= 20; j++) {
