+++ Project Euler/src/Q6/src/Q6.h
