+++ Project Euler/src/Q7/src/Q7.h
