+++ Project Euler/src/Q9/src/Q9.h
